Day 15

+++ src/main/java/uk/co/eandrews/util/Vector2.java
@@ -9,19 +9,19 @@ public record Vector2(long x, long y) {
 
     public static final Vector2 ZERO = new Vector2(0, 0);
 
-    public Vector2 up(final int step) {
+    public Vector2 up(final long step) {
         return toBuilder().x(x).y(y + step).build();
     }
 
-    public Vector2 down(final int step) {
+    public Vector2 down(final long step) {
         return toBuilder().x(x).y(y - step).build();
     }
 
-    public Vector2 left(final int step) {
+    public Vector2 left(final long step) {
         return toBuilder().x(x - step).y(y).build();
     }
 
-    public Vector2 right(final int step) {
+    public Vector2 right(final long step) {
         return toBuilder().x(x + step).y(y).build();
     }
 }
